+++ packages/typespec-go/test/http-specs/type/model/inheritance/singlediscgroup/singlediscriminator_client_test.go
@@ -87,12 +87,35 @@ func TestSingleDiscriminatorClientGetWrongDiscriminator(t *testing.T) {
 	}, resp.BirdClassification)
 }
 
+func TestSingleDiscriminatorClientGetNoSubtypesModel(t *testing.T) {
+	client, err := singlediscgroup.NewSingleDiscriminatorClientWithNoCredential("http://localhost:3000", nil)
+	require.NoError(t, err)
+	resp, err := client.GetNoSubtypesModel(context.Background(), nil)
+	require.NoError(t, err)
+	require.NotNil(t, resp.FishClassification)
+	require.Equal(t, &singlediscgroup.Fish{
+		Kind: to.Ptr("salmon"),
+		Size: to.Ptr[int32](10),
+	}, resp.FishClassification)
+}
+
 func TestSingleDiscriminatorClientPutModel(t *testing.T) {
 	client, err := singlediscgroup.NewSingleDiscriminatorClientWithNoCredential("http://localhost:3000", nil)
 	require.NoError(t, err)
 	resp, err := client.PutModel(context.Background(), &singlediscgroup.Sparrow{
 		Kind:     to.Ptr("sparrow"),
 		Wingspan: to.Ptr[int32](1),
+	}, nil)
+	require.NoError(t, err)
+	require.Zero(t, resp)
+}
+
+func TestSingleDiscriminatorClientPutNoSubtypesModel(t *testing.T) {
+	client, err := singlediscgroup.NewSingleDiscriminatorClientWithNoCredential("http://localhost:3000", nil)
+	require.NoError(t, err)
+	resp, err := client.PutNoSubtypesModel(context.Background(), &singlediscgroup.Fish{
+		Kind: to.Ptr("salmon"),
+		Size: to.Ptr[int32](10),
 	}, nil)
 	require.NoError(t, err)
 	require.Zero(t, resp)